Use merged loki libs in grafana-agent
